+++ src/heuristic_functions.h
@@ -3,7 +3,7 @@
 
 #include "solver_steps.h"
 
-int dr_heuristic(cube_t* cube, ptable_data_t* p_data);
-int htr_heuristic(cube_t* cube, ptable_data_t* p_data);
+size_t dr_heuristic(cube_t* cube, ptable_data_t* p_data);
+size_t htr_heuristic(cube_t* cube, ptable_data_t* p_data);
 
 #endif /* HEURISTIC_FUNCIONS_H */
